@MLnick 2017-04-24
Interesting - I was working on something very similar - a rough draft of it is in a [branch](https://github.com/mlnick/spark/tree/SPARK-13857-als-parity-v3).