@MLnick 2017-04-26
Could you post updated performance numbers?

I think we can do the same optimization in `ml` version. We could do it in a follow up PR perhaps?