@mpjlu 2017-04-26
Another case
3 workers， each 40 cores, each 196G memory, each 1 executor.
Data Size: user 480,000, item 17,000 
recommendProductsForUsers with blockSize 4096 is about 34s