@mpjlu 2017-04-26
Thanks very much @MLnick . 
I am doing more test about mllib solution. When it is solid enough, then we can submit a follow up PR for ML optimization. How do you think about it?