@MLnick 2017-04-27
@mpjlu yeah we can do the ML version in a follow up PR that is ok (I can help if needed).