@mpjlu 2017-04-28
Hi @MLnick, I will be on vacation next week.
If you have time to create an ML optimization follow up PR. I am ok. Otherwise, I will submit the follow up PR after my 1 week vacation.
Thanks very much.