@MLnick 2017-04-28
@mpjlu I should be able to put something together for the ML API equivalent to this code - in fact I already have a version close to ready to go :)

Thanks for this PR and major performance improvement! I will take a final pass over this shortly.