@MLnick 2017-05-03
This LGTM - it would be nice to have this and #17845 in 2.2 (especially since the `ml` recommend all methods are new in 2.2 and this gives a massive performance gain).

cc @srowen @jkbradley