@MLnick 2017-05-03
@mpjlu I did just think about the hand-written `dot` for larger `rank` (say 100-200 which is not so common as say 10-50 range for ALS, but still can occur). Did you compare performance for that range of `rank` at all?