@mpjlu 2017-05-04
@MLnick , I did not test rank in range 100 to 200. I am on vacation and my colleague is doing this test, I will update this result soon. 
Thanks.