@MLnick 2017-05-04
Ok - I don't expect it would make a substantial difference for dot unless
the vector size was very large.

Let's check it but we could proceed to merge this in the meantime.


On Thu, 4 May 2017 at 15:58, Meng, Peng <notifications@github.com> wrote:

> @MLnick <https://github.com/MLnick> , I did not test rank in range 100 to
> 200. I am on vacation and my colleague is doing this test, I will update
> this result soon.
> Thanks.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/apache/spark/pull/17742#issuecomment-299192898>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AA_SB_R-VfrZJtchobPNOAneGs-H613Oks5r2dmVgaJpZM4NF_tV>
> .
>