@mpjlu 2017-05-05
@MLnick , for rank is 200, the F2jblas dot is about the same performance  as handwritten dot. 
Both of them is much better than native dot.