@jkbradley 2017-05-08
Thanks for doing this!  The logic looks reasonable to me.

One comment: I'm not a fan of custom BLAS implementations scattered throughout MLlib.  I'm OK if people want to merge this now, but could you please follow up by putting the dot as a private API in BLAS.scala and adding unit tests?

@mpjlu You can remove the "[ML]" tag from the PR title since @MLnick is addressing that API separately.