@MLnick 2017-05-09
@jkbradley good point. In fact thinking about it again now, it's best to simply use `F2jBLAS` since the performance is the same as the hand-written version. That's actually what we already do internally for MLlib's `BLAS.dot` for `Vector`.

I've created [SPARK-20677](https://issues.apache.org/jira/browse/SPARK-20677) to do that places once they're merged.