@MLnick 2017-05-09
Merged to master/branch-2.2.

Thanks @mpjlu for the awesome performance improvement!