@mpjlu 2017-05-10
Thanks @mengxr , glad to meet you here. I am Meng Peng.
I have tested different blockSize, see https://issues.apache.org/jira/browse/SPARK-20443
I will test the other methods you mentioned.