@mpjlu 2017-05-10
I don't think we should use BLAS 3 here, because no matter use output or not here, you need a big buffer to save the BLAS result. That still cause GC problem. 
I also want to test: build a handwritten native gemm,  which not returns C (C=A*B), but only return the topK elements of each row.  This maybe better performance than current solution.