@mpjlu 2017-05-10
Thanks, I will do some test based on BLAS 3.