@mpjlu 2017-05-10
** The most optimized version would be doing a quickselect on each row and select top k.
** An easy-to-implement version would be:
I test both of the methods, the best performance is about 50% this PR.
Welcome to do more test.